renames app.routing.js to app.config.js, creates homeRoutes.js and index.js inside the home directory
