+++ app/app.config.js
@@ -1,0 +1,9 @@
+routing.$inject = ['$urlRouterProvider', '$locationProvider'];
+
+export default function routing($urlRouterProvider, $locationProvider) {
+   $locationProvider.html5Mode({
+  		enabled: true,
+  		requireBase: false
+   });
+   $urlRouterProvider.otherwise('/');
+}
