+++ app/app.module.js
@@ -1,11 +1,11 @@
 import angular from 'angular';
 import uiRouter from 'angular-ui-router';
 
-import routing from './app.routing.js';
+import routes from './app.config.js';
 import ex from './components/ex';
 
 angular.module('feeder', [uiRouter, 'ngMaterial', ex])
-	.config(routing);
+	.config(routes);
